Create For Loop: Create a new list from  the existing list with a condition
In the previous exercise, you adapted the data structure you were looping over. Remember, it's also possible to build a new data structure altogether inside your for loop.

Your plant manager approaches you and asks for a report on all failures that are available in the logs list. Instead of the entire list, she is only interested in the failures. Get to work to generate what she asks for!

Just a tip before you get to it: If you have a list of lists a and want to add a list b to it, you can use c(a, list(b)).

INSTRUCTIONS
100XP
Initialize an empty list, failures using the list() function without arguments.
Finish the for loop such that each log entry that indicates a failure is added to failures.
Display the structure of the failures list that results.