Create selecting text from list of characters
Infer gender from name (2)
In the previous exercise, it appeared that the title Mr. may not cover all men on board. Instead of manually going through all titles that appear in the Name column of titanic, we can write a clever gsub() command that extracts the title part.

The pattern we'll need is the following:

"^.*, (.*?)\\..*$"
With ^ and $ we signify the start and end of the string. Next, we have two .* parts in there: wildcards for the last name and first names. With , (.*?)\\. we use a similar pattern as before, but the parentheses allow us to re-use whatever is matched inside the parentheses in our replacement.

INSTRUCTIONS
100XP
Fill in the pattern into the gsub() call to create the titles vector. The "\\1" part tells R to replace the entire string with whatever is matched inside the parentheses.
Call unique() on titles to get an overview of all different titles that are found in the name vector. Simply print out the result.